Decouple RepoParser from git
This is the first step in adding support for other SCMS by splitting the git
implementation out of repo_parser.RepoParser.
